+++ src/app/locales/en/components/content-spaces-providers.js
@@ -6,4 +6,5 @@ export default {
     header: 'Add support',
     text: 'To store data, the space must be supported by at least one provider. ',
   },
+  ceaseSupportItem: 'Cease support',
 };