+++ CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
-
+* VFS-5907 Added cease Oneprovider support action in supporting Oneproviders space view
 * VFS-5875 Fixed navigation to GUI settings of a cluster and optimized rendering of sidebars
 * VFS-5861 Fixed corrupted rendering membership paths with "more" or "forbidden" blocks
 * VFS-1891 Added privacy policy and cookie consent notification